requirements.txt
aiosignal==1.3.1
annotated-types==0.6.0
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-apscheduler==0.6.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
frozenlist==1.4.1
google-api-core==2.15.0
google-api-python-client==2.113.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
multidict==6.0.4
oauthlib==3.2.2
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.7.1
pydantic_core==2.18.2
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
